feat: multi-version Revit support (2024/2025/2026)
- csproj: dual TFM (net48 + net8.0-windows) with RevitYear build param,
  per-year output dirs, DefineConstants (REVIT2024/2025/2026)
- element IDs: int -> long across all DTOs, CLI options, interface,
  client, controllers, and operations (64-bit ElementId since Revit 2024)
- ToCliElementId: removed checked int cast, returns long directly
- install.ps1: auto-detects available years, deploys per-year add-in
  binaries and manifests, install.json stores revitYears array
- uninstall.ps1: removes manifests for all installed years
- release.yml: builds add-in separately for 2024/2025/2026,
  packages addin/2024, addin/2025, addin/2026 in ZIP

Build: dotnet build -f net48 -p:RevitYear=2024
       dotnet build -f net8.0-windows -p:RevitYear=2025
       dotnet build -f net8.0-windows -p:RevitYear=2026

+++ shared/RevitCli.Shared/AuditResult.cs
@@ -27,5 +27,5 @@ public class AuditIssue
     public string Message { get; set; } = "";
 
     [JsonPropertyName("elementId")]
-    public int? ElementId { get; set; }
+    public long? ElementId { get; set; }
 }
